fix(nika-cli): shrink under 15k and teach leftover first-run doors
Descend catalog, explain, and the help postcard into nika-cli-host so
nika-cli production LOC sits under the 15k wall. Split check tests and
shorten nika try's example() under the file/fn ratchets.

C11 UX-2 B07 B08: the postcard glosses permits as a field in the file
and isolation as a note; `nika permits` teaches explain / infer-permits
instead of clap-dumping.

C01: the pause card names `--answer key=true`; yes/y and no/n coerce
to the same boolean as true/false.

C13: `nika check --fix` wraps a bare exec string as command: + shell:
and a root `nika --fix` names `nika check --fix`.

TTY ghost `mcp:` tools print NIKA-INVOKE-001, matching JSON.
