+++ crates/nika-cli/src/verbs/mod.rs
@@ -11,11 +11,9 @@
 //! testable as a library call.
 
 pub mod arm;
-pub mod catalog;
 pub mod check;
 pub mod context;
 pub mod examples;
-pub mod explain;
 pub mod explain_file;
 pub mod fix;
 pub mod graph;
@@ -34,7 +32,7 @@ pub mod session;
 pub mod sign;
 pub mod test;
 pub mod tools;
-pub use nika_cli_host::{doctor, probe, welcome, wire};
+pub use nika_cli_host::{catalog, doctor, explain, probe, welcome, wire};
 // The trace-reading plane descended to `nika-trace` 2026-08-11 (the 15k
 // prod-LOC wall · D-2026-07-09-N1 one unit, two members · the ADR-110
 // cli-host precedent) — re-exported at the historical verbs:: paths so
